refactor: ChessBoard의 Column 수정
